requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
frozenlist==1.3.3
gunicorn==20.1.0
heroku3==5.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaggle==1.5.13
loguru==0.6.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.0
pandas==1.5.3
pinecone-client==2.2.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
text-unidecode==1.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
